refactor: Unify project packages (#130)
* refactor(accounts): Use a single file for all domain errors

* refactor(accounts): Use a single file for all requests schemas

* refactor(accounts): Remove http prefix from controllers and routes files

* refactor(courses): Update files names

* chore(labs): Track responses folder

Add an empty .gitkeep file to track the folder

* refactor(lang): Update files names and track requests folder

* refactor(rubrics): Use a single file for all domain errors

* refactor(rubrics): Remove http prefix from controllers and routes files

* refactor(session): Use a single file for all dtos

* refactor(session): Use a single file for all domain errors

* refactor(session): Remove http prefix from controllers and routes files

* refactor(session): Use a single file for all requests and track responses folder

* refactor: Replace snake case imports with camel case

Use golang naming convention
