+++ src/accounts/infrastructure/http/routes.go
@@ -3,7 +3,7 @@ package infrastructure
 import (
 	"github.com/UPB-Code-Labs/main-api/src/accounts/application"
 	"github.com/UPB-Code-Labs/main-api/src/accounts/infrastructure/implementations"
-	shared_infrastructure "github.com/UPB-Code-Labs/main-api/src/shared/infrastructure"
+	sharedInfrastructure "github.com/UPB-Code-Labs/main-api/src/shared/infrastructure"
 	"github.com/gin-gonic/gin"
 )
 
@@ -22,25 +22,25 @@ func StartAccountsRoutes(g *gin.RouterGroup) {
 	accountsGroup.POST("/students", controller.HandleRegisterStudent)
 	accountsGroup.GET(
 		"/students",
-		shared_infrastructure.WithAuthenticationMiddleware(),
-		shared_infrastructure.WithAuthorizationMiddleware([]string{"teacher"}),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthorizationMiddleware([]string{"teacher"}),
 		controller.HandleSearchStudents,
 	)
 	accountsGroup.POST(
 		"/admins",
-		shared_infrastructure.WithAuthenticationMiddleware(),
-		shared_infrastructure.WithAuthorizationMiddleware([]string{"admin"}),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthorizationMiddleware([]string{"admin"}),
 		controller.HandleRegisterAdmin,
 	)
 	accountsGroup.GET(
 		"/admins",
-		shared_infrastructure.WithAuthenticationMiddleware(),
-		shared_infrastructure.WithAuthorizationMiddleware([]string{"admin"}),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthorizationMiddleware([]string{"admin"}),
 		controller.HandleGetAdmins,
 	)
 	accountsGroup.POST("/teachers",
-		shared_infrastructure.WithAuthenticationMiddleware(),
-		shared_infrastructure.WithAuthorizationMiddleware([]string{"admin"}),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthorizationMiddleware([]string{"admin"}),
 		controller.HandleRegisterTeacher,
 	)
 }
