+++ src/courses/domain/errors/courses_errors.go
