+++ src/languages/infrastructure/http/routes.go
@@ -3,7 +3,7 @@ package http
 import (
 	"github.com/UPB-Code-Labs/main-api/src/languages/application"
 	"github.com/UPB-Code-Labs/main-api/src/languages/infrastructure/implementations"
-	shared_infrastructure "github.com/UPB-Code-Labs/main-api/src/shared/infrastructure"
+	sharedInfrastructure "github.com/UPB-Code-Labs/main-api/src/shared/infrastructure"
 	"github.com/gin-gonic/gin"
 )
 
@@ -20,14 +20,14 @@ func StartLanguagesRoutes(g *gin.RouterGroup) {
 
 	langGroup.GET(
 		"",
-		shared_infrastructure.WithAuthenticationMiddleware(),
-		shared_infrastructure.WithAuthorizationMiddleware([]string{"teacher", "student"}),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthorizationMiddleware([]string{"teacher", "student"}),
 		controllers.HandleGetLanguages,
 	)
 	langGroup.GET(
 		"/:language_uuid/template",
-		shared_infrastructure.WithAuthenticationMiddleware(),
-		shared_infrastructure.WithAuthorizationMiddleware([]string{"teacher", "student"}),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthorizationMiddleware([]string{"teacher", "student"}),
 		controllers.HandleDownloadLanguageTemplate,
 	)
 }
