+++ src/rubrics/domain/errors/criteria_errors.go
