+++ src/rubrics/domain/errors/objectives_errors.go
