--- src/session/domain/dtos/login_dto.go
@@ -1,6 +1,0 @@
-package dtos
-
-type LoginDTO struct {
-	Email    string
-	Password string
-}
