+++ src/session/domain/errors/session_errors.go
