+++ src/session/infrastructure/http/routes.go
@@ -2,7 +2,7 @@ package http
 
 import (
 	accounts_impl "github.com/UPB-Code-Labs/main-api/src/accounts/infrastructure/implementations"
-	shared_infrastructure "github.com/UPB-Code-Labs/main-api/src/shared/infrastructure"
+	sharedInfrastructure "github.com/UPB-Code-Labs/main-api/src/shared/infrastructure"
 
 	"github.com/UPB-Code-Labs/main-api/src/session/application"
 	"github.com/gin-gonic/gin"
@@ -14,7 +14,7 @@ func StartSessionRoutes(g *gin.RouterGroup) {
 	useCases := application.SessionUseCases{
 		AccountsRepository: accounts_impl.GetAccountsPgRepository(),
 		PasswordHasher:     accounts_impl.GetArgon2PasswordsHasher(),
-		TokenHandler:       shared_infrastructure.GetJwtTokenHandler(),
+		TokenHandler:       sharedInfrastructure.GetJwtTokenHandler(),
 	}
 
 	controllers := &SessionControllers{
@@ -25,13 +25,13 @@ func StartSessionRoutes(g *gin.RouterGroup) {
 
 	sessionGroup.DELETE(
 		"/logout",
-		shared_infrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
 		controllers.HandleLogout,
 	)
 
 	sessionGroup.GET(
 		"/whoami",
-		shared_infrastructure.WithAuthenticationMiddleware(),
+		sharedInfrastructure.WithAuthenticationMiddleware(),
 		controllers.HandleWhoAmI,
 	)
 }
